ai_agent/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
python-dotenv==1.0.1
requests==2.32.0
python-multipart==0.0.10
google-generativeai==0.8.3
beautifulsoup4==4.12.3
lxml==5.3.0
wikipedia-api==0.6.0
fandom-py==0.2.1
psycopg2-binary==2.9.9 